Unity support

+++ Program.cs
@@ -4,6 +4,7 @@ using System.Threading;
 using System.Diagnostics;
 using System.IO;
 using VitaDefiler.Modules;
+using VitaDefiler.PSM;
 
 namespace VitaDefiler
 {
@@ -11,24 +12,51 @@ namespace VitaDefiler
     {
         static readonly Type[] Mods = {typeof(Code), typeof(General), typeof(Memory), typeof(FileIO), typeof(Scripting)};
 
+        public static bool exitAfterInstall = false;
+
         static void Main(string[] args)
         {
+            int scriptIndex = 0;
             bool enablegui = true;
+            string package = null;
 
+            foreach (string arg in args)
+            {
+                switch (arg)
+                {
+                    case "-nodisp":
+                        ++scriptIndex;
+                        enablegui = false;
+                        break;
+
+                    case "-install":
+                        scriptIndex += 2;
+                        package = args[0];
+                        exitAfterInstall = true;
+                        break;
+                }
+            }
+
+#if !USE_UNITY
             if (args.Length < 1)
             {
                 Console.Error.WriteLine("usage: VitaDefiler.exe package [-nodisp] [script args]\n    package is path to PSM package\n    nodisp starts client without logging to screen\n    script is the script to run\n    args are arguments for the script");
                 return;
             }
-            if (!File.Exists(args[0]))
+
+            if (string.IsNullOrEmpty(package))
+            {
+                package = args[0];
+                ++scriptIndex;
+            }
+#endif
+
+            if (!string.IsNullOrEmpty(package) && !File.Exists(package))
             {
                 Console.Error.WriteLine("cannot find package file");
                 return;
             }
-            if (args.Length >= 2 && args[1] == "-nodisp")
-            {
-                enablegui = false;
-            }
+
 #if USE_APP_KEY
             if (!File.Exists(args[1]))
             {
@@ -37,16 +65,22 @@ namespace VitaDefiler
             }
 #endif
 
-            // kill PSM
-            Process[] potential = Process.GetProcessesByName("PsmDevice");
-            foreach (Process process in potential)
+            if (Environment.OSVersion.VersionString.Contains("Microsoft Windows"))
             {
-                Console.WriteLine("Killing PsmDevice process {0}", process.Id);
-                process.Kill();
+                // kill PSM
+                Process[] potential = Process.GetProcesses();
+                foreach (Process process in potential)
+                {
+                    if (process.ProcessName.StartsWith("PsmDevice") || process.ProcessName.StartsWith("PsmDeviceUnity"))
+                    {
+                        Console.WriteLine("Killing PsmDevice process {0}", process.Id);
+                        process.Kill();
+                    }
+                }
             }
 
             // set environment variables
-            Environment.SetEnvironmentVariable("SCE_PSM_SDK", Path.Combine(Environment.CurrentDirectory, "support"));
+            Environment.SetEnvironmentVariable("SCE_PSM_SDK", Path.Combine(Environment.CurrentDirectory, "support/psm"));
 
             // initialize the modules
             List<IModule> mods = new List<IModule>();
@@ -65,59 +99,41 @@ namespace VitaDefiler
             }
 
             // set up usb
-            USB usb = new USB(args[0], null);
-            ManualResetEvent doneinit = new ManualResetEvent(false);
-            string host = string.Empty;
-            int port = 0;
-            usb.Connect((text) =>
-            {
-                if (text.StartsWith("XXVCMDXX:"))
-                {
-#if DEBUG
-                    Console.Error.WriteLine("[Vita] {0}", text);
+            Exploit exploit;
+            string host;
+            int port;
+            
+#if USE_UNITY
+                ExploitFinder.CreateFromWireless(package, out exploit, out host, out port);
+#else
+                ExploitFinder.CreateFromUSB(package, out exploit, out host, out port);
 #endif
-                    string[] cmd = text.Trim().Split(':');
-                    switch (cmd[1])
-                    {
-                        case "IP":
-                            host = cmd[2];
-                            port = Int32.Parse(cmd[3]);
-                            Console.Error.WriteLine("Found Vita network at {0}:{1}", host, port);
-                            break;
-                        case "DONE":
-                            Console.Error.WriteLine("Vita done initializing");
-                            doneinit.Set();
-                            break;
-                        default:
-                            Console.Error.WriteLine("Unrecognized startup command");
-                            break;
-                    }
-                }
-                else
-                {
-                    Console.Error.WriteLine("[Vita] {0}", text);
-                }
-            });
-            Console.Error.WriteLine("Waiting for app to finish launching...");
-            doneinit.WaitOne();
 
+#if !NO_EXPLOIT
             uint images_hash_ptr;
             uint[] funcs = new uint[5];
             uint logline_func;
             uint libkernel_anchor;
             Console.Error.WriteLine("Defeating ASLR...");
-            usb.DefeatASLR(out images_hash_ptr, out funcs[0], out funcs[1], out funcs[2], out funcs[3], out funcs[4], out libkernel_anchor);
-#if !NO_ESCALATE_PRIVILEGES
+            exploit.DefeatASLR(out images_hash_ptr, out funcs[0], out funcs[1], out funcs[2], out funcs[3], out funcs[4], out libkernel_anchor);
             // exploit vita
+
             Console.Error.WriteLine("Escalating privileges...");
-            usb.EscalatePrivilege(images_hash_ptr);
+            exploit.EscalatePrivilege(images_hash_ptr);
+#endif
+
+#if USE_UNITY
+            exploit.ResumeVM(); // The network listener is already listening in Unity.
+#else
+            exploit.StartNetworkListener();
+            Console.Error.WriteLine("Vita exploited.");
+#endif
+
+
             //Thread tt = new Thread(() =>
             //{
-                usb.StartNetworkListener();
-                Console.Error.WriteLine("Vita exploited.");
             //});
                 //tt.Start();
-#endif
 
             // set up network
             Network net = new Network();
@@ -128,7 +144,7 @@ namespace VitaDefiler
             else
             {
                 Console.Error.WriteLine("Failed to create net listener. Exiting.");
-                usb.Disconnect();
+                exploit.Disconnect();
                 return;
             }
 
@@ -140,16 +156,19 @@ namespace VitaDefiler
                 Console.Error.WriteLine("Enabling display output");
                 net.RunCommand(Command.EnableGUI, out resp);
             }
-
+            
+#if !NO_EXPLOIT
             // pass in function pointers
             if (net.RunCommand(Command.SetFuncPtrs, funcs, out resp) == Command.Error)
             {
                 Console.Error.WriteLine("ERROR setting function pointers!");
             }
+#endif
 
             // set up RPC context
-            Device dev = new Device(usb, net);
-
+            Device dev = new Device(exploit, net);
+            
+#if !NO_EXPLOIT
             // get logger
             net.RunCommand(Command.GetLogger, out resp);
             logline_func = BitConverter.ToUInt32(resp, 0);
@@ -162,24 +181,15 @@ namespace VitaDefiler
             dev.CreateLocal("pss_code_mem_flush_icache", funcs[4]);
             dev.CreateLocal("logline", logline_func);
             dev.CreateLocal("libkernel_anchor", libkernel_anchor);
+#endif
 
             // run script if needed
-            if ((!enablegui && args.Length >= 3) || (enablegui && args.Length >= 2))
+            if (args.Length > scriptIndex)
             {
-                string script;
-                string[] scriptargs;
-                if (enablegui)
-                {
-                    script = args[1];
-                    scriptargs = new string[args.Length - 2];
-                    Array.Copy(args, 2, scriptargs, 0, args.Length - 2);
-                }
-                else
-                {
-                    script = args[2];
-                    scriptargs = new string[args.Length - 3];
-                    Array.Copy(args, 3, scriptargs, 0, args.Length - 3);
-                }
+                string script = args[scriptIndex];
+                string[] scriptargs = new string[args.Length - scriptIndex - 1];
+                Array.Copy(args, scriptIndex + 1, scriptargs, 0, args.Length - scriptIndex - 1);
+
                 scripting.ParseScript(dev, script, scriptargs);
             }
 
@@ -282,7 +292,7 @@ namespace VitaDefiler
             }
 
             // cleanup
-            usb.Disconnect();
+            exploit.Disconnect();
         }
     }
 }
